finished creating the mechanism for hiding objects
I created a method in the Map class called revealObject which reveals
changes the object's boolean isHidden to false. There is also a method
called revealAll which loops through the whole grid and reveals
everything. We will use this to show Rodriguez that there are elements
stored in the map.